requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
decorator==4.2.1
defusedxml==0.5.0
Django==1.11.12
django-allauth==0.33.0
django-appconf==1.0.2
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap3==9.0.0
django-cities-light==3.5.0
django-compressor==2.2
django-des==2.2.0
django-extensions==2.0.7
django-filter==1.1.0
django-rest-auth==0.9.3
django-solo==1.1.2
django-uuidfield==0.5.0
django-webpack-loader==0.6.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djmail==1.0.1
graphviz==0.8.2
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
kombu==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
msgpack-python==0.4.8
oauthlib==2.0.6
olefile==0.44
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
progressbar2==3.38.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
pycparser==2.18
pydot==1.2.4
pydot-ng==1.0.0
Pygments==2.2.0
pygraphviz==1.3.1
PyJWT==1.6.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.3.0
python3-openid==3.1.0
pytz==2018.3
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
Twisted==17.9.0
txaio==2.8.2
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
zope.interface==4.4.3
